fix: ChatInterfaceの責務の分離
