+++ frontend/src/features/ChatInterface/hooks/useChatStreaming.ts
@@ -1,0 +1,471 @@
+import { useAtom, useSetAtom } from "jotai";
+import { useCallback, useMemo, useRef } from "react";
+import { useTranslation } from "react-i18next";
+
+import {
+	addMessageWithIdAtom,
+	updateMessageAtom,
+} from "../../../store/chatAtoms";
+import { currentLanguageAtom } from "../../../store/languageAtoms";
+
+import { useSentiment } from "../../../hooks/useSentiment";
+import { useStreamingTTS } from "../../../hooks/useStreamingTTS";
+import { useTextToSpeech } from "../../../hooks/useTextToSpeech";
+
+import {
+	type ConversationMessage,
+	buildPrompt,
+	generateTextNonStreaming,
+	generateTextStream,
+} from "../../../services/llmService";
+
+import type { VRMWrapperHandle } from "../../VRM/VRMWrapper/VRMWrapper";
+
+/**
+ * ChatStreamingフック用のオプション型定義
+ */
+export interface UseChatStreamingOptions {
+	/** VRMアバターを制御するためのwrapper参照 */
+	vrmWrapperRef?: React.RefObject<VRMWrapperHandle | null>;
+}
+
+/**
+ * ストリーミング状態を管理する型定義
+ */
+export interface StreamingState {
+	/** LLMからのレスポンス待機状態 */
+	isLoading: boolean;
+	/** テキスト生成処理中状態 */
+	isGenerating: boolean;
+	/** テキスト表示アニメーション実行中状態 */
+	isAnimating: boolean;
+}
+
+/**
+ * ChatStreamingフックの戻り値型定義
+ */
+export interface ChatStreamingResult {
+	/** 現在のストリーミング状態 */
+	state: StreamingState;
+	/** AI応答生成を開始する関数 */
+	generateResponse: (
+		query: string,
+		conversationHistory: ConversationMessage[],
+		isStreamingMode: boolean,
+	) => Promise<void>;
+	/** 現在の生成処理を停止する関数 */
+	stopGeneration: () => void;
+	/** テキストアニメーション表示関数 */
+	animateText: (speed?: number) => void;
+}
+
+/**
+ * チャットインターフェース用のストリーミング機能を管理するカスタムフック
+ *
+ * 主な機能:
+ * - LLMからのストリーミング/非ストリーミングレスポンス処理
+ * - リアルタイムテキストアニメーション表示
+ * - TTSストリーミング音声合成との連携
+ * - センチメント分析によるVRM表情制御
+ * - メッセージ間のバッファ分離と安全な状態管理
+ *
+ * @param options フック設定オプション
+ * @returns ストリーミング制御インターフェース
+ */
+export const useChatStreaming = ({
+	vrmWrapperRef,
+}: UseChatStreamingOptions): ChatStreamingResult => {
+	// ===== 状態管理 =====
+	const [currentLanguage] = useAtom(currentLanguageAtom);
+	const addMessageWithId = useSetAtom(addMessageWithIdAtom);
+	const updateMessage = useSetAtom(updateMessageAtom);
+	const { t } = useTranslation("chat");
+
+	// ===== 内部状態ref =====
+	/** ストリーミングテキストを一時保存するバッファ */
+	const streamBuffer = useRef("");
+	/** テキストアニメーション実行フラグ */
+	const isAnimating = useRef(false);
+	/** ローディング状態フラグ */
+	const isLoading = useRef(false);
+	/** テキスト生成処理中フラグ */
+	const isGenerating = useRef(false);
+	/** 現在処理中のメッセージID */
+	const lastMessageId = useRef<number | null>(null);
+	/** アニメーション表示用の現在のテキスト */
+	const currentDisplayText = useRef("");
+	/** API呼び出し中断用のコントローラー */
+	const abortRef = useRef<AbortController | null>(null);
+
+	// ===== バッファリング機能用の状態 =====
+	/** TTSストリーミング用チャンクバッファ */
+	const chunkBufferRef = useRef<string>("");
+	/** バッファ送信タイマー */
+	const bufferTimerRef = useRef<NodeJS.Timeout | null>(null);
+
+	// ===== 外部フック =====
+	/** ストリーミングTTS機能 */
+	const streamingTTSHook = useStreamingTTS({
+		vrmWrapperRef,
+	});
+
+	/** streamingTTSをメモ化してlintエラーを回避 */
+	const streamingTTS = useMemo(
+		() => ({
+			addChunk: streamingTTSHook.addChunk,
+			finalize: streamingTTSHook.finalize,
+			stopStreaming: streamingTTSHook.stopStreaming,
+			clearQueue: streamingTTSHook.clearQueue,
+			isReady: streamingTTSHook.isReady,
+		}),
+		[
+			streamingTTSHook.addChunk,
+			streamingTTSHook.finalize,
+			streamingTTSHook.stopStreaming,
+			streamingTTSHook.clearQueue,
+			streamingTTSHook.isReady,
+		],
+	);
+
+	const streamingTTSState = streamingTTSHook.state;
+
+	/** 従来のTTS機能 */
+	const { speak, stop: stopLegacyTTS } = useTextToSpeech({
+		vrmWrapperRef,
+	});
+
+	/** センチメント分析機能 */
+	const { analyzeSentiment } = useSentiment({
+		enabled: true,
+		enableDebugLogging: true,
+	});
+
+	// ===== ユーティリティ関数 =====
+	/**
+	 * 一意のメッセージIDを生成
+	 * タイムスタンプとカウンターを組み合わせて重複を防ぐ
+	 */
+	const messageIdCounter = useRef(0);
+	const createId = useCallback(() => {
+		messageIdCounter.current += 1;
+		return Date.now() * 1000 + messageIdCounter.current;
+	}, []);
+
+	/**
+	 * バッファリング機能付きチャンク送信関数
+	 * 連続したチャンクをバッファし、50ms後にバッチ送信することで
+	 * TTSストリーミングの効率を向上させる
+	 *
+	 * @param chunk 送信するテキストチャンク
+	 */
+	const bufferedAddChunk = useCallback(
+		(chunk: string) => {
+			chunkBufferRef.current += chunk;
+
+			// 既存のタイマーをクリアして新しいタイマーを設定
+			if (bufferTimerRef.current) {
+				clearTimeout(bufferTimerRef.current);
+			}
+
+			// 50ms後にバッファの内容をまとめて送信
+			bufferTimerRef.current = setTimeout(() => {
+				if (chunkBufferRef.current) {
+					streamingTTS.addChunk(chunkBufferRef.current);
+					chunkBufferRef.current = "";
+				}
+			}, 50);
+		},
+		[streamingTTS.addChunk],
+	);
+
+	/**
+	 * テキストをタイピング風にアニメーション表示する関数
+	 * 1文字ずつ順次表示してリアルタイム感を演出
+	 *
+	 * @param speed アニメーション速度（ms）、デフォルト30ms
+	 */
+	const animateText = useCallback(
+		(speed = 30) => {
+			// アニメーション停止条件のチェック
+			if (!isAnimating.current || streamBuffer.current.length === 0) {
+				isAnimating.current = false;
+				return;
+			}
+
+			// バッファから1文字取り出して表示テキストに追加
+			const char = streamBuffer.current.substring(0, 1);
+			streamBuffer.current = streamBuffer.current.substring(1);
+			currentDisplayText.current += char;
+
+			// メッセージIDが有効で、コンテンツがある場合のみ更新
+			// これにより空のメッセージや無効なIDによる更新を防ぐ
+			if (lastMessageId.current && currentDisplayText.current) {
+				updateMessage({
+					id: lastMessageId.current,
+					updates: { text: currentDisplayText.current },
+				});
+			}
+
+			// 次の文字のアニメーションをスケジュール
+			setTimeout(() => requestAnimationFrame(() => animateText(speed)), speed);
+		},
+		[updateMessage],
+	);
+
+	/**
+	 * 全ての音声再生を停止する関数
+	 * ストリーミングTTSと従来のTTSの両方を停止
+	 */
+	const stopAllAudio = useCallback(() => {
+		stopLegacyTTS();
+		streamingTTS.stopStreaming();
+	}, [stopLegacyTTS, streamingTTS.stopStreaming]);
+
+	/**
+	 * 全てのバッファと状態をクリアする関数
+	 * メッセージ間の汚染を防ぐため、新しいメッセージ開始時に呼び出す
+	 */
+	const clearBuffers = useCallback(() => {
+		// バッファタイマーをクリア
+		if (bufferTimerRef.current) {
+			clearTimeout(bufferTimerRef.current);
+			bufferTimerRef.current = null;
+		}
+		// 全バッファとアニメーション状態をリセット
+		chunkBufferRef.current = "";
+		streamBuffer.current = "";
+		currentDisplayText.current = "";
+		isAnimating.current = false;
+	}, []);
+
+	/**
+	 * 現在の生成処理を完全に停止する関数
+	 * API呼び出し、音声再生、アニメーションを全て停止
+	 */
+	const stopGeneration = useCallback(() => {
+		abortRef.current?.abort(); // API呼び出し中断
+		stopAllAudio(); // 音声停止
+		isLoading.current = false; // ローディング状態解除
+		isGenerating.current = false; // 生成状態解除
+		clearBuffers(); // バッファクリア
+	}, [stopAllAudio, clearBuffers]);
+
+	/**
+	 * AI応答生成のメイン関数
+	 * ストリーミング/非ストリーミングモードに対応し、
+	 * TTS、センチメント分析、アニメーション表示を統合的に管理
+	 *
+	 * @param query ユーザーの質問テキスト
+	 * @param conversationHistory 会話履歴
+	 * @param isStreamingMode ストリーミングモードの有効/無効
+	 */
+	const generateResponse = useCallback(
+		async (
+			query: string,
+			conversationHistory: ConversationMessage[],
+			isStreamingMode: boolean,
+		) => {
+			// 既に生成中の場合は処理をスキップ
+			if (isGenerating.current) return;
+
+			// 前のメッセージ処理を完全に停止
+			// 複数のメッセージが同時に処理されることを防ぐ
+			if (isAnimating.current) {
+				isAnimating.current = false;
+			}
+
+			// 生成状態フラグを設定
+			isGenerating.current = true;
+			isLoading.current = true;
+
+			// API呼び出し用のAbortControllerを準備
+			abortRef.current?.abort(); // 前の呼び出しがあれば中断
+			const controller = new AbortController();
+			abortRef.current = controller;
+
+			// 新しいAIメッセージIDを生成
+			const aiMessageId = createId();
+			lastMessageId.current = aiMessageId;
+
+			// バッファを完全にクリアして前のメッセージとの混入を防ぐ
+			clearBuffers();
+
+			// 空のAIメッセージをチャット履歴に追加
+			addMessageWithId({
+				id: aiMessageId,
+				text: "",
+				isUser: false,
+				isStreaming: isStreamingMode,
+			});
+
+			// TTSキューをクリア
+			streamingTTS.clearQueue();
+
+			try {
+				// ユーザーの質問をプロンプトに変換
+				const payloadQuery = buildPrompt(query);
+
+				if (isStreamingMode) {
+					// ===== ストリーミングモード =====
+					let accumulatedText = ""; // 蓄積されたレスポンステキスト
+
+					await generateTextStream(
+						payloadQuery,
+						conversationHistory,
+						controller.signal,
+						(chunk) => {
+							if (chunk.type === "content" && chunk.content) {
+								// チャンクは増分コンテンツとして扱う
+								// 以前は累積テキストとの差分計算をしていたが、
+								// これが原因でメッセージ間での内容混入が発生していた
+								const incrementalText = chunk.content;
+
+								if (incrementalText) {
+									// テキストを蓄積
+									accumulatedText += incrementalText;
+									streamBuffer.current += incrementalText;
+
+									// アニメーションが停止中の場合は開始
+									if (!isAnimating.current) {
+										isAnimating.current = true;
+										animateText(); // デフォルト速度（30ms）を使用
+									}
+
+									// TTSストリーミング用にチャンクをバッファリング
+									bufferedAddChunk(incrementalText);
+								}
+							} else if (chunk.type === "done") {
+								// ストリーミング完了時の処理
+								streamingTTS.finalize(); // TTSストリーミング終了
+								analyzeSentiment(accumulatedText); // センチメント分析
+
+								// アニメーション完了を待ってから状態を更新
+								const waitForAnimation = () => {
+									if (isAnimating.current || streamBuffer.current.length > 0) {
+										// まだアニメーション中の場合は100ms後に再チェック
+										setTimeout(waitForAnimation, 100);
+										return;
+									}
+									// アニメーション完了時の最終処理
+									updateMessage({
+										id: aiMessageId,
+										updates: { isStreaming: false },
+									});
+									isLoading.current = false;
+									isGenerating.current = false;
+								};
+								waitForAnimation();
+
+								// フォールバック機能: ストリーミングTTSが開始されない場合の対処
+								setTimeout(() => {
+									const { isPlaying, isGenerating, queue } = streamingTTSState;
+									if (
+										!isPlaying &&
+										!isGenerating &&
+										queue.length === 0 &&
+										accumulatedText.trim()
+									) {
+										console.warn(
+											"Streaming TTS did not start. Using legacy TTS.",
+										);
+										speak(accumulatedText); // 従来のTTSにフォールバック
+									}
+								}, 2000);
+							}
+						},
+						"/api/llm/query",
+						currentLanguage,
+					);
+				} else {
+					// ===== 非ストリーミングモード =====
+					const response = await generateTextNonStreaming(
+						payloadQuery,
+						conversationHistory,
+						controller.signal,
+						currentLanguage,
+					);
+
+					// レスポンスを文字単位でアニメーション表示
+					// バッファをクリアしてから新しいレスポンスを設定
+					streamBuffer.current = response;
+					currentDisplayText.current = "";
+
+					// アニメーション開始
+					if (!isAnimating.current) {
+						isAnimating.current = true;
+						animateText(20); // 非ストリーミングモードでは少し速めに表示
+					}
+
+					// 非ストリーミングモードでは通常のTTSを使用
+					speak(response);
+
+					// センチメント分析を実行
+					analyzeSentiment(response);
+
+					// アニメーション完了を待ってからローディング状態を解除
+					const waitForAnimation = () => {
+						if (isAnimating.current || streamBuffer.current.length > 0) {
+							setTimeout(waitForAnimation, 100);
+							return;
+						}
+						updateMessage({
+							id: aiMessageId,
+							updates: { isStreaming: false },
+						});
+						isLoading.current = false;
+						isGenerating.current = false;
+					};
+					waitForAnimation();
+				}
+			} catch (err) {
+				// エラー処理: 全ての処理を停止してエラーメッセージを表示
+				stopAllAudio();
+				isLoading.current = false;
+				isGenerating.current = false;
+				clearBuffers();
+
+				if (err instanceof Error && err.name === "AbortError") {
+					// ユーザーによる中断の場合
+					updateMessage({
+						id: aiMessageId,
+						updates: { text: t("generationStopped"), isStreaming: false },
+					});
+				} else {
+					// その他のエラーの場合
+					console.error("Text generation error:", err);
+					updateMessage({
+						id: aiMessageId,
+						updates: { text: t("errorGeneratingResponse"), isStreaming: false },
+					});
+				}
+			}
+		},
+		[
+			createId,
+			addMessageWithId,
+			streamingTTS,
+			bufferedAddChunk,
+			animateText,
+			analyzeSentiment,
+			updateMessage,
+			speak,
+			currentLanguage,
+			t,
+			stopAllAudio,
+			clearBuffers,
+			streamingTTSState,
+		],
+	);
+
+	// ===== フック戻り値 =====
+	return {
+		state: {
+			isLoading: isLoading.current,
+			isGenerating: isGenerating.current,
+			isAnimating: isAnimating.current,
+		},
+		generateResponse,
+		stopGeneration,
+		animateText,
+	};
+};
